Rename cache to reduce confusing naming semantics
