+++ lib/ramble/ramble/util/install_cache.py
@@ -1,0 +1,19 @@
+# Copyright 2022 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 <LICENSE-APACHE or
+# https://www.apache.org/licenses/LICENSE-2.0> or the MIT license
+# <LICENSE-MIT or https://opensource.org/licenses/MIT>, at your
+# option. This file may not be copied, modified, or distributed
+# except according to those terms.
+
+"""Provide data structures to assist with cacheing operations"""
+
+class SetCache:
+    def __init__(self):
+        self.store = set()
+
+    def add(self, tupl):
+        self.store.add(tupl)
+
+    def contains(self, tupl):
+        return (tupl in self.store)
